requirements.txt
flask==3.1.1
google-genai==1.73.1
flask-cors==5.0.1
python-dotenv==1.1.0
